fix: resolve all compilation errors from incomplete refactoring
Fix import paths for classes moved during r17 refactoring:
- ChapterType: bare import → eu.kanade.tachiyomi.animesource.model
- IndexedSegment: SeekBar.IndexedSegment → controls.components
- VideoTrack: PlayerViewModel.VideoTrack → model.VideoTrack (4 files)
- openContentFd/getFileName: util.system → ui.player.PlayerUtils

Fix PlayerActivity references to refactored APIs:
- Replace episodeListManager.updateNavigationState() with delegating method
- Replace direct isLoadingTracks.update() with setLoadingTracks()

Add updateNavigationState() to PlayerViewModel.
